refactor(logic): make setup town assignment importable (Refs #1879)
Extract town tile ranking into game_setup_town_tile_ranking.dart and
convert game_setup_helpers_towns from a part to a standalone library.
game_setup_helpers.dart re-exports town APIs for existing imports.

Remove two stale function-size allowlist rows with wrong symbol names.
Document module layout in SPEC/program/game-setup-pipeline.md.

Add unit tests for town tile ranking helpers.

Made-with: Cursor
