+++ packages/colonizethis_logic/lib/src/setup/game_setup_helpers.dart
@@ -1,17 +1,16 @@
-import 'dart:math';
-
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
 import '../constants.dart';
 import '../world/naval.dart';
 import '../world/ship_instance_allocate.dart';
-import 'capital_choice.dart';
 import 'game_setup_context.dart';
 import 'game_setup_create.dart';
+import 'game_setup_helpers_towns.dart';
 import 'province_name_fallback.dart';
 
-part 'game_setup_helpers_towns.dart';
+export 'game_setup_helpers_towns.dart';
+
 part 'game_setup_helpers_naming.dart';
 part 'game_setup_helpers_bootstrap.dart';
 
@@ -127,66 +126,6 @@ List<String> provinceIdsFromTopology(MapTopology topology) {
       .where((n) => n.type == TopologyNodeType.province)
       .map((n) => n.id)
       .toList();
-}
-
-({int x, int y}) provinceTownCentroidFromTileKeys(List<String> tiles) {
-  final c = roundedCentroidFromTileKeys(tiles);
-  if (c != null) return c;
-  final xy = parseTileKeyCellXY(tiles.first);
-  return (x: xy?.$1 ?? 0, y: xy?.$2 ?? 0);
-}
-
-int compareTownTileCandidates(
-  String a,
-  String b, {
-  required int centroidX,
-  required int centroidY,
-  required Map<String, int> bfsFromCapital,
-}) {
-  final da = tileDistanceSquaredToCentroid(a, centroidX: centroidX, centroidY: centroidY);
-  final db = tileDistanceSquaredToCentroid(b, centroidX: centroidX, centroidY: centroidY);
-  if (da != db) return da.compareTo(db);
-  final ba = bfsDistanceOrUnreachable(a, bfsFromCapital);
-  final bb = bfsDistanceOrUnreachable(b, bfsFromCapital);
-  if (ba != bb) return ba.compareTo(bb);
-  return a.compareTo(b);
-}
-
-int tileDistanceSquaredToCentroid(
-  String tileKey, {
-  required int centroidX,
-  required int centroidY,
-}) {
-  final xy = parseTileKeyCellXY(tileKey);
-  if (xy == null) return 1 << 30;
-  final dx = xy.$1 - centroidX;
-  final dy = xy.$2 - centroidY;
-  return dx * dx + dy * dy;
-}
-
-int bfsDistanceOrUnreachable(String tileKey, Map<String, int> bfsFromCapital) {
-  const unreachable = 999999;
-  return bfsFromCapital[tileKey] ?? unreachable;
-}
-
-String pickTownTileByCentroidAndBfs({
-  required List<String> candidates,
-  required int centroidX,
-  required int centroidY,
-  required Map<String, int> bfsFromCapital,
-}) {
-  return candidates.reduce(
-    (best, candidate) => compareTownTileCandidates(
-          candidate,
-          best,
-          centroidX: centroidX,
-          centroidY: centroidY,
-          bfsFromCapital: bfsFromCapital,
-        ) <
-        0
-        ? candidate
-        : best,
-  );
 }
 
 /// Re-runs §7d province town assignment after the caller mutates provinces or maps.
